Add files via upload
This program uses an artificial recurrent neural network called Long Short Term Memory (LSTM) to predict the closing stock price of GME using past 60 day stock price. LSTM is an artificial recurrent neural network architecture used in the field of deep learning. They are widely used for sequence prediction problems and have proven to be extremely effective. the reason why they are so effective is because they are able to store past information that is important and forget the information that is not important. 

NOTE: This is an experiment and if you'd like to use it for actual stock or crypto predictions, please note that the stock prediction is with 35% error rate. Although I used the same code to predict ETH prices and was able to get the results that I wanted in real world, that was a stroke of luck. 35% error rate is very high in LSTM prediction models. the ideal error rate should be within 1 to 5%.